requirements.txt
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.1
bleach==3.1.0
bootstrap-admin==0.4.4
bs4==0.0.1
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
defusedxml==0.6.0
Django==2.2.8
django-allauth==0.40.0
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-haystack==2.8.1
django-imagekit==4.0.2
django-redis==4.10.0
django-uuslug==1.2.0
djangorestframework==3.10.3
gunicorn==19.9.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itemadapter==0.1.0
jieba==0.39
lxml==4.5.1
Markdown==3.1.1
oauthlib==3.1.0
olefile==0.46
parsel==1.6.0
pilkit==2.0
Pillow==7.1.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.2
queuelib==1.5.0
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
Scrapy==2.2.0
scrapy-djangoitem==1.1.1
service-identity==18.1.0
six==1.12.0
soupsieve==2.0.1
sqlparse==0.3.0
text-unidecode==1.3
Twisted
urllib3==1.25.3
user-agent==0.1.9
w3lib==1.22.0
webencodings==0.5.1
Whoosh==2.7.4
zope.interface==5.1.0
